+++ Assets/Basegame/Audio.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 965e640b69f5bda46bfd894d7c2c8572
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
