+++ Assets/Basegame/Scripts/PlayerCollision.cs
@@ -6,12 +6,19 @@ public class PlayerCollision : MonoBehaviour {
 
     public PlayerMovement move;
     public Rigidbody rb;
+    public AudioSource audio;
+
+    void Start()
+    {
+        audio = GetComponent<AudioSource>();
+    }
 
     private void OnCollisionEnter(Collision collision)
     {
         if (collision.collider.tag == "Obstacle")
         {
             move.enabled = false;
+            audio.Play();
             FindObjectOfType<GameManager>().EndGame();
 
         }
