--- Assets/ShowScore.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e416d064390a90746a2027d81b62f41d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
